add images for mobile more height add script to manage the carrousel slider correct some styles in the reset.css add styles for the slide delete some atribute on the HTML and
