+++ scripts/sriptManager.js
@@ -1,1 +1,4 @@
 import './menuTogle.js'
+import './slider.js'
+
+
