+++ styles/reset.css
@@ -64,7 +64,7 @@ video,
 canvas,
 svg {
   display: block;
-  max-width: 100%;
+  width: 100%;
   height: auto;
 }
