+++ styles/slideStyles.css
@@ -1,6 +1,6 @@
 .slider {
-  max-height: 420px;
   position: relative;
+  /* overflow:hidden */
 }
 
 .slider__controls {
@@ -14,12 +14,20 @@
     background-color: rgb(204, 188, 204, 0.9);
     border: 0;
     padding: 8px;
+    z-index: 1;
   }
 }
 
-.slide__one,
-.slide__two {
-  display: none;
+.slider__wrapper {
+  /* overflow: hidden; */
+  display: flex;
+  width: 100%;
+  transition: transform 0.5s ease-in-out;
+}
+
+.slider__slide {
+  flex: 0 0 100%;
+  min-width: 100%;
 }
 
 .slider__pagination {
@@ -37,6 +45,7 @@
     height: 10px;
     width: 10px;
   }
+  
   .slider__dot--active {
     background-color: var(--orange);
   }
