+++ styles/styles.css
@@ -56,7 +56,7 @@
   opacity: 1;
   transform: translateY(0);
   visibility: visible;
-  z-index: 1;
+  z-index: 2;
 }
 
 .menu__item-link {
